pyproject.toml
[project]
name = "ursina-proteins"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ursina (>=7.0.0,<8.0.0)",
    "biopython (>=1.85,<2.0)"
]

[tool.poetry]
packages = [{include = "ursina_proteins", from = "src"}]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pytest = "^8.3.5"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
